refactor(torghut): split candidate overlay ownership
