+++ services/torghut/tests/test_explicit_module_facade_exports.py
@@ -26,13 +26,21 @@ EXPLICIT_IMPORT_SURFACES = (
     "app.trading.discovery.candidate_spec_base_profiles_primary",
     "app.trading.discovery.candidate_spec_base_profiles_secondary",
     "app.trading.discovery.candidate_spec_common",
+    "app.trading.discovery.candidate_spec_execution_research_overlays",
+    "app.trading.discovery.candidate_spec_family_selection",
     "app.trading.discovery.candidate_spec_feedback_profiles",
     "app.trading.discovery.candidate_spec_hpairs_profiles",
+    "app.trading.discovery.candidate_spec_hypothesis_policy",
+    "app.trading.discovery.candidate_spec_mechanism_overlay_state",
+    "app.trading.discovery.candidate_spec_mechanism_overlays",
+    "app.trading.discovery.candidate_spec_microstructure_overlays",
     "app.trading.discovery.candidate_spec_portfolio_profiles_primary",
     "app.trading.discovery.candidate_spec_portfolio_profiles_secondary",
     "app.trading.discovery.candidate_spec_profile_constants",
     "app.trading.discovery.candidate_spec_profile_variants",
     "app.trading.discovery.candidate_spec_profiles",
+    "app.trading.discovery.candidate_spec_replay_validation_overlays",
+    "app.trading.discovery.candidate_spec_stat_arb_impact_overlays",
     "app.trading.discovery.evidence_bundles",
     "app.trading.discovery.fast_replay",
     "app.trading.discovery.microstructure_prefilter",
